Add interactive mode for track input and enhance track processing logic

+++ sisthatsrisky/spoti.py
@@ -7,13 +7,31 @@ import os
 import subprocess
 import json
 import time
+import sys
 from bs4 import BeautifulSoup
 import ytmusicapi
 
 yt = ytmusicapi.YTMusic()
-f = open("tracks.txt", "r")
-tracks = f.readlines()
-f.close()
+
+# Check for interactive mode
+interactive_mode = "-i" in sys.argv
+
+if interactive_mode:
+    tracks = []
+    print("Interactive mode enabled. Enter search queries one at a time.")
+    print("Format: Artist - Title (e.g. The Beatles - Hey Jude)")
+    print("Type 'exit' to quit.\n")
+    while True:
+        query = input("Enter search query (e.g. Artist - Title): ").strip()
+        if query.lower() == "exit":
+            break
+        if query:
+            tracks.append(query)
+else:
+    f = open("tracks.txt", "r")
+    tracks = f.readlines()
+    f.close()
+
 tracknum = len(tracks)
 currtrack = 0
 
@@ -54,23 +72,32 @@ for surl in tracks:
         print(f"Processed {currtrack} tracks. Taking a 60 second break...")
         time.sleep(60)
 
-    try:
-        trackid = surl.replace("https://open.spotify.com/track/", "").split("?")[0]
-    except:
-        trackid = surl.replace("https://open.spotify.com/track/", "")
-    if not trackid in downloaded:
-        print("Track " + str(currtrack) + " of " + str(tracknum) + " (" + trackid.split("\n")[0] + ")")
-        if not "http" in surl:
-            surl = "https://open.spotify.com/track/" + surl
-        print("Downloading...")
-        with urllib.request.urlopen(surl) as response:
-            r = response.read().decode()
-        soup = BeautifulSoup(r, "lxml")
-        artist = soup.find("meta", {"name": "music:musician_description"})["content"].split(",")[0]
-        title = soup.find("meta", {"property": "og:title"})["content"]
-        filename = artist.replace(" ", "-").replace("/", "-") + "_" + title.replace(" ", "-").replace("/", "-")
-        filename = slugify(filename) + ".mp3"
-        if not os.path.exists(filename):
+    # Check if this is a custom entry
+    if surl.startswith("#custom:"):
+        trackid = surl.replace("#custom:", "").strip()
+        query = trackid
+        artist = trackid.split(" - ")[0] if " - " in trackid else "Unknown"
+        title = trackid.split(" - ")[1] if " - " in trackid else trackid
+        year = ""
+        albumtitle = ""
+        track = ""
+        album_maxtracks = ""
+        cover = None
+    else:
+        try:
+            trackid = surl.replace("https://open.spotify.com/track/", "").split("?")[0]
+        except:
+            trackid = surl.replace("https://open.spotify.com/track/", "")
+        if not trackid in downloaded:
+            print("Track " + str(currtrack) + " of " + str(tracknum) + " (" + trackid.split("\n")[0] + ")")
+            if not "http" in surl:
+                surl = "https://open.spotify.com/track/" + surl
+            print("Downloading...")
+            with urllib.request.urlopen(surl) as response:
+                r = response.read().decode()
+            soup = BeautifulSoup(r, "lxml")
+            artist = soup.find("meta", {"name": "music:musician_description"})["content"].split(",")[0]
+            title = soup.find("meta", {"property": "og:title"})["content"]
             data = json.loads(soup.find("script", {"type": "application/ld+json"}).contents[0])
             cover = soup.find("meta", {"property": "og:image"})["content"]
             year = data["datePublished"]
@@ -81,24 +108,42 @@ for surl in tracks:
             soup2 = BeautifulSoup(r2, "lxml")
             albumtitle = soup2.find("meta", {"property": "og:title"})["content"]
             album_maxtracks = soup2.find("meta", {"property": "og:description"})["content"].split(" · ")[3].split(" ")[0]
+        
+        query = artist + " - " + title
+
+    filename = artist.replace(" ", "-").replace("/", "-") + "_" + title.replace(" ", "-").replace("/", "-")
+    filename = slugify(filename) + ".mp3"
+    if not os.path.exists(filename):
+        if cover:
             os.system("wget -O audio.jpg \"" + cover + "\"")
-            print(title)
-            query = artist + " - " + title
-            print(query)
-            try:
-                videoId = yt.search(query, "songs")[0]["videoId"]
-            except:
-                videoId = None
-            if not videoId == None:
-                cmd = ["yt-dlp", "--no-continue", "--add-metadata", "-x", "--prefer-ffmpeg", "--extract-audio", "-v", "--audio-format", "mp3", "--output", "audio.%(ext)s", "https://youtu.be/"+videoId]
-            else:
-                cmd = ["yt-dlp", "--no-continue", "--add-metadata", "-v", "--prefer-ffmpeg", "--extract-audio", "-v", "--audio-format", "mp3", "--output", "audio.%(ext)s", "ytsearch:\"" + query + "\"", "--no-playlist"]
-            subprocess.Popen(cmd, shell=False).wait()
-            print("Converting...")
-            subprocess.Popen(["lame", "-b", "320", "--ti", "audio.jpg", "--ta", artist, "--tt", title, "--ty", year, "--tl", albumtitle, "--tn", track+"/"+album_maxtracks, "audio.mp3", filename], shell=False).wait()
-            os.system("rm audio.mp3")
+        print(title)
+        print(query)
+        try:
+            videoId = yt.search(query, "songs")[0]["videoId"]
+        except:
+            videoId = None
+        if not videoId == None:
+            cmd = ["yt-dlp", "--no-continue", "--add-metadata", "-x", "--prefer-ffmpeg", "--extract-audio", "-v", "--audio-format", "mp3", "--output", "audio.%(ext)s", "https://youtu.be/"+videoId]
+        else:
+            cmd = ["yt-dlp", "--no-continue", "--add-metadata", "-v", "--prefer-ffmpeg", "--extract-audio", "-v", "--audio-format", "mp3", "--output", "audio.%(ext)s", "ytsearch:\"" + query + "\"", "--no-playlist"]
+        subprocess.Popen(cmd, shell=False).wait()
+        print("Converting...")
+        lame_cmd = ["lame", "-b", "320"]
+        if cover:
+            lame_cmd.extend(["--ti", "audio.jpg"])
+        lame_cmd.extend(["--ta", artist, "--tt", title])
+        if year:
+            lame_cmd.extend(["--ty", year])
+        if albumtitle:
+            lame_cmd.extend(["--tl", albumtitle])
+        if track and album_maxtracks:
+            lame_cmd.extend(["--tn", track+"/"+album_maxtracks])
+        lame_cmd.extend(["audio.mp3", filename])
+        subprocess.Popen(lame_cmd, shell=False).wait()
+        os.system("rm audio.mp3")
+        if cover:
             os.system("rm audio.jpg")
-            print("Done!")
-        f = open("downloaded.txt", "a+")
-        f.write(trackid + "\n")
-        f.close()
+        print("Done!")
+    f = open("downloaded.txt", "a+")
+    f.write(trackid + "\n")
+    f.close()
